requirements.txt
aniso8601==9.0.1
ascii-magic==2.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.1
coverage==7.6.9
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
docutils==0.21.2
executing==2.0.1
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
fonttools==4.55.0
git-filter-repo==2.47.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
gunicorn==23.0.0
httplib2==0.22.0
identify==2.6.3
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jwt==1.3.1
kiwisolver==1.4.7
limits==3.13.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
marshmallow-sqlalchemy==1.1.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.5
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.0
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==12.6.0
rich-cli==1.8.0
rich-rst==1.3.1
rpds-py==0.21.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
textual==0.1.18
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.0.4
whitenoise==6.8.2
wrapt==1.16.0
